travel_backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
mongoengine==0.27.0
django-rest-framework-mongoengine==3.4.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
python-dotenv==1.0.0
pymongo==4.6.0
djongo==1.3.6